chore: make sending sms follow the flow of letter and rec letter.
